+++ examples/kcp/server.go
@@ -13,7 +13,6 @@ func main() {
 	// server peer
 	srv := drpc.NewEndpoint(drpc.EndpointConfig{
 		Network:     "kcp",
-		CountTime:   true,
 		ListenPort:  9090,
 		PrintDetail: true,
 	})