+++ examples/quick_start/server.go
@@ -8,7 +8,6 @@ import (
 func main() {
 	// 创建一个rpc服务
 	svr := drpc.NewEndpoint(drpc.EndpointConfig{
-		CountTime:   true,
 		LocalIP:     "127.0.0.1",
 		ListenPort:  9091,
 		PrintDetail: true,