+++ examples/redis/server.go
@@ -17,7 +17,6 @@ func main() {
 	go graceful.GraceSignal()
 	db = gmap.New(true)
 	svr := drpc.NewEndpoint(drpc.EndpointConfig{
-		CountTime:   true,
 		LocalIP:     "127.0.0.1",
 		ListenPort:  9091,
 		PrintDetail: true,